+++ docs/spec/src/index.ts
@@ -2,9 +2,10 @@ import { IntegrationSpecConfig } from '@jupiterone/integration-sdk-core';
 
 import { IntegrationConfig } from '../../../src/config';
 import { accountSpec } from './account';
+import { hostSpec } from './host';
 import { roleSpec } from './role';
 import { userSpec } from './user';
 
 export const invocationConfig: IntegrationSpecConfig<IntegrationConfig> = {
-  integrationSteps: [...userSpec, ...accountSpec, ...roleSpec],
+  integrationSteps: [...userSpec, ...accountSpec, ...roleSpec, ...hostSpec],
 };