+++ src/steps/constants.ts
@@ -10,32 +10,44 @@ export const Steps = {
   USERS: 'fetch-users',
   ACCOUNT: 'fetch-account',
   ROLES: 'fetch-roles',
+  HOSTS: 'fetch-hosts',
   BUILD_ACCOUNT_USER_RELATIONSHIPS: 'build-account-user-relationships',
   BUILD_ACCOUNT_ROLE_RELATIONSHIPS: 'build-account-role-relationships',
+  BUILD_ACCOUNT_HOST_RELATIONSHIPS: 'build-account-host-relationships',
   BUILD_USER_ROLE_RELATIONSHIPS: 'build-user-role-relationships',
 };
 
-export const Entities: Record<'USER' | 'ACCOUNT' | 'ROLE', StepEntityMetadata> =
-  {
-    USER: {
-      resourceName: 'User',
-      _type: 'datadog_user',
-      _class: ['User'],
-    },
-    ACCOUNT: {
-      resourceName: 'Account',
-      _type: 'datadog_account',
-      _class: ['Account'],
-    },
-    ROLE: {
-      resourceName: 'Role',
-      _type: 'datadog_role',
-      _class: ['AccessRole'],
-    },
-  };
+export const Entities: Record<
+  'USER' | 'ACCOUNT' | 'ROLE' | 'HOST',
+  StepEntityMetadata
+> = {
+  USER: {
+    resourceName: 'User',
+    _type: 'datadog_user',
+    _class: ['User'],
+  },
+  ACCOUNT: {
+    resourceName: 'Account',
+    _type: 'datadog_account',
+    _class: ['Account'],
+  },
+  ROLE: {
+    resourceName: 'Role',
+    _type: 'datadog_role',
+    _class: ['AccessRole'],
+  },
+  HOST: {
+    resourceName: 'Host',
+    _type: 'datadog_host',
+    _class: ['Host'],
+  },
+};
 
 export const Relationships: Record<
-  'ACCOUNT_HAS_USER' | 'ACCOUNT_HAS_ROLE' | 'USER_ASSIGNED_ROLE',
+  | 'ACCOUNT_HAS_USER'
+  | 'ACCOUNT_HAS_ROLE'
+  | 'ACCOUNT_HAS_HOST'
+  | 'USER_ASSIGNED_ROLE',
   StepRelationshipMetadata
 > = {
   ACCOUNT_HAS_USER: {
@@ -49,6 +61,12 @@ export const Relationships: Record<
     _class: RelationshipClass.HAS,
     sourceType: Entities.ACCOUNT._type,
     targetType: Entities.ROLE._type,
+  },
+  ACCOUNT_HAS_HOST: {
+    _type: 'datadog_account_has_host',
+    _class: RelationshipClass.HAS,
+    sourceType: Entities.ACCOUNT._type,
+    targetType: Entities.HOST._type,
   },
   USER_ASSIGNED_ROLE: {
     _type: 'datadog_user_assigned_role',